+++ atlasdb-commons/src/main/java/com/palantir/async/initializer/Callback.java
@@ -16,6 +16,7 @@
 
 package com.palantir.async.initializer;
 
+import java.util.List;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 
@@ -78,6 +79,45 @@ public abstract class Callback<R> {
 
         @Override
         public void cleanup(R resource, Throwable initException) {
+        }
+    }
+
+    /**
+     * A CallChain executes a list of callbacks in sequence.
+     *
+     * Note that a given callback is executed only once all previous callbacks in the chain have successfully executed.
+     * Also, callbacks are retried independently. In other words, given a CallChain of two callbacks C1 and C2:
+     *
+     * <ul>
+     *     <li>
+     *         C1 is executed first. If it fails, C1's cleanup will be invoked, and then C1 will be retried.
+     *     </li>
+     *     <li>
+     *         C2 is then executed. If it fails, C2's cleanup will be invoked and then C2 will be retried.
+     *         (Note that we will not re-invoke C1's cleanup, nor will we execute it again.)
+     *         Once C2 succeeds, the CallChain as a whole will be treated as complete.
+     *     </li>
+     * </ul>
+     *
+     * Please see the examples in CallChainTest for more detail.
+     */
+    public static class CallChain<T> extends Callback<T> {
+        private final List<Callback<T>> callbacks;
+
+        public CallChain(List<Callback<T>> callbacks) {
+            this.callbacks = callbacks;
+        }
+
+        @Override
+        public void init(T resource) {
+            callbacks.forEach(callback -> callback.runWithRetry(resource));
+        }
+
+        @Override
+        public void cleanup(T resource, Throwable cleanupException) {
+            // Rethrows, because each callback's runWithRetry is responsible for cleanup of any resources needed
+            // to be cleaned up for that task.
+            throw Throwables.rewrapAndThrowUncheckedException(cleanupException);
         }
     }
 }